Refactor: Transform from mock to fully working Kubernetes integration
Major Changes:
- Replace simulated data with real Kubernetes API integration
- Add kubernetes Python client library (>=28.0.0) to dependencies
- Implement K8sClient class for cluster communication
- Add support for real metrics: pod status, replicas, node pressure, resource usage
- Convert main.py from symlink to standalone file
- Add namespace parameter to CLI (-n/--namespace)
- Create config.yaml for customizable settings
- Update README with accurate usage instructions and troubleshooting

Implementation Details:
- get_k8s_data() now fetches actual cluster metrics via k8s API
- Supports both kubeconfig (~/.kube/config) and in-cluster configuration
- Changed USE_KUBE_CONFIG default to "true" for easier local development
- Added comprehensive error handling for API failures
- CPU/memory metrics derived from resource requests (metrics-server optional)

Cleanup:
- Remove duplicate src/sonify_k8s.py file
- Remove empty sound WAV files (sounds/ directory)
- Fix missing # in color codes for replicas metric
- Remove built-in modules (logging, typing) from requirements.txt

This transforms the application from a demo/mock into a production-ready
Kubernetes monitoring tool that connects to real clusters and sonifies
actual cluster health metrics.
