+++ requirements.txt
@@ -1,5 +1,3 @@
-logging
-typing
 random2
 pylint
 pytest
@@ -9,3 +7,5 @@ numpy>=1.20.0
 click>=8.0.0
 mido>=1.2.10
 python-rtmidi>=1.4.9
+kubernetes>=28.0.0
+PyYAML>=6.0.0
